requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.28
werkzeug==3.0.1
jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.9.0
eventlet==0.35.2 